feat: mig-smoke-ready-deletion-batches-001 - Migrate smoke-delete-13 wrong-layer verify contracts
Move verify-tier Makefile and CI script contract proofs from the smoke catch-all
into tests/functional/observability/verification and retire the smoke-owned file.
