+++ docs/internal/baselines/functional-undocumented-tests.json
@@ -1187,82 +1187,6 @@
       "name": "TestStatelessExecutionSmoke_LoadedConfigDrivesExecution"
     },
     {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestBackendCoverageAliasesSmoke_RedirectToIndependentLanes"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestBackendVerificationLaneScriptSmoke_PreservesFailureExitAndLog"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestBackendVerificationLaneScriptSmoke_UsesCanonicalOwnedCommandAndCapturesLog"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestConcurrentUIVerificationLanesScriptSmoke_DoesNotWaitForDetachedLogHandle"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestConcurrentUIVerificationLanesScriptSmoke_FailureReportsExactLaneRerun"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestConcurrentUIVerificationLanesScriptSmoke_RunsBothOwnedLanesConcurrently"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestLongTestsCommandSmoke_FailureReportsExactSpecialtyLaneRerun"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestUICoverageCommandSmoke_RunsPackageCoverageThenReplayCheck"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestUIPackageCoverageCommandSmoke_InvokesPackageOwnedCoverageScript"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyCompatibilityAliasSmoke_RedirectsToCanonicalPRTier"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyExtendedCommandSmoke_UsesOnlyExplicitLongSuitesAfterPRTier"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyFastCommandSmoke_ContractFailureStopsLaterSuites"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyFastCommandSmoke_FailureReportsOwnedSuiteAndRerunCommand"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyFastCommandSmoke_UsesOnlyShortOwnedSuites"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyPRCommandSmoke_FailureReportsExactLaneRerun"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyPRCommandSmoke_UsesRequiredLanesOnce"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyPRInferenceCommandSmoke_FailureReportsOwnedRerunCommand"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyPRInferenceCommandSmoke_RunsSingleNamedRegressionOnly"
-    },
-    {
-      "file": "smoke/verify_fast_command_smoke_test.go",
-      "name": "TestVerifyPRInferenceCommandSmoke_StaysOutsideRequiredPRAndExtendedTiers"
-    },
-    {
       "file": "work/recordings/recordings_read_test.go",
       "name": "TestGetWorkFromRecordingsRootUsesRecordingsServiceRoot"
     },
