+++ docs/internal/baselines/package-structure-baseline.json
@@ -1329,11 +1329,6 @@
     },
     {
       "rule": "functional-test-missing-subsection",
-      "filePath": "tests/functional/smoke/verify_fast_command_smoke_test.go",
-      "target": "smoke"
-    },
-    {
-      "rule": "functional-test-missing-subsection",
       "filePath": "tests/functional/workflow/code_review_loop_long_test.go",
       "target": "workflow"
     },
