+++ tests/functional/observability/verification/doc.go
@@ -1,0 +1,3 @@
+// Package verification owns Make-contract and CI script proofs for verify-*
+// tier wiring, wrapper scripts, and lane ordering.
+package verification
